Moved NVIDIA *util.h files to External and adjusted build
