+++ External/TBTK/External/NVIDIACuda/cusolverMg_utils.h
@@ -33,7 +33,7 @@
 #include <stdexcept>
 
 //TODO remove TBTK/ when done
-#include "TBTK/cusolver_utils.h"
+#include "TBTK/Extern/NVIDIACuda/cusolver_utils.h"
 
 #ifndef IDX2F
 #define IDX2F(i, j, lda) ((((j)-1) * (static_cast<size_t>(lda))) + ((i)-1))
